refactor: created a pages folder.
Moved home.js and BodyComposition.js to the new pages folder.
